pyproject.toml
[tool.poetry]
name = "obras"
version = "0.1.0"
description = ""
authors = ["Luiz-aluisio <2019aluisio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
asgiref = "3.7.2"
django = "4.2.6"
django-countries = "7.5.1"
django-cpf-cnpj = "1.0.0"
python-decouple = "3.8"
pyuca = "1.2"
sqlparse = "0.4.4"
typing-extensions = "4.8.0"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
iniconfig = "2.0.0"
packaging = "23.2"
pluggy = "1.3.0"
pytest = "7.4.3"
pytest-django = "4.5.2"
coverage = "7.3.2"
pytest-cov = "4.1.0"
taskipy = "^1.12.0"
blue = "^0.9.1"
ruff = "^0.1.3"
isort = "^5.12.0"
model-bakery = "^1.17.0"
factory-boy = "^3.3.0"

[tool.ruff]
line-length = 79
exclude = [".venv", "migrations"]

[tool.isort]
profile = "black"
line_length = 79
extend_skip = ["migrations"]

[tool.pytest.ini_options]
pythonpath = "."

[tool.blue]
extend-exclude = "(migrations/)"

[tool.taskipy.tasks]
lint = "ruff . && blue --check . --diff"
format = "blue . && isort ."
run = "python manage.py runserver"
pre_test = "task lint"
test = "pytest -s -x --cov=arte -vv"
post_test = "coverage html"
db-migrate = 'python manage.py  migrate'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
